Custom Routing Started, Contentful Hooked In & Markdown Processing
Custom routing is using Navigo - and requires switchstatements to
transition between component pages.

There are 2 new pages - Home and LandingPage which is no longer part of
app so it's separate from the routing.

Markdown processing is required because that's how it's set in
contentful. Not sure if I like the method currently used but it works.