Style Alterations Made To Encompass Desktop
Mobile looked great. THe screen is full of content, but on desktop
it just wasn't working. The image was not good but that's a content
thing, but the text was altered. Just one font is in use site wide.

A screen wide image is used at the top of the post to add a little
something for desktop purposes.